--- Assets/Scripts/Asteroid.cs
@@ -1,38 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Asteroid : MonoBehaviour
-{
-    [SerializeField]
-    private float _rotateSpeed = 19.0f;
-    [SerializeField]
-    private GameObject _explosionPrefab;
-    private SpawnManager _spawnManager;
-
-
-    void Start()
-    {
-        _spawnManager = GameObject.Find("Spawn_Manager").GetComponent<SpawnManager>();
-    }
-
-    void Update()
-    {
-        transform.Rotate(Vector3.forward * _rotateSpeed * Time.deltaTime);
-    }
-
-
-    public void OnTriggerEnter2D(Collider2D other)
-    {
-        if (other.tag == "Laser")
-        {
-            Instantiate(_explosionPrefab, transform.position, Quaternion.identity);
-            Destroy(other.gameObject);
-            _spawnManager.StartSpawning();
-            Destroy(this.gameObject, 0.25f);
-        }
-
-
-    }
-
-}
