--- Assets/Scripts/Explosion.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Explosion : MonoBehaviour
-{
-
-    void Start()
-    {
-        Destroy(this.gameObject, 3.0f);
-    }
-}
