--- Assets/Scripts/GameManager.cs
@@ -1,24 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class GameManager : MonoBehaviour
-{
-    [SerializeField]
-    private bool _isGameOver;
-
-    void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.R) && _isGameOver == true)
-        {
-            SceneManager.LoadScene(1); //Current Game Scene
-        }
-    }
-
-    public void GameOver()
-    {
-        _isGameOver = true;
-    }
-
-}
