--- Assets/Scripts/Laser.cs
@@ -1,26 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Laser : MonoBehaviour
-{
-    [SerializeField]
-    private float _laserSpeed = 8f;
-
-
-    void Update()
-    {
-
-        Vector3 laserDir = Vector3.up * _laserSpeed * Time.deltaTime;
-        transform.Translate(laserDir);
-
-        if (transform.position.y > 7.5f)
-        {
-            if (transform.parent != null)
-            {
-                Destroy(transform.parent.gameObject);
-            }
-            Destroy(this.gameObject);
-        }
-    }
-}
