+++ Assets/Scripts/Misc/Laser.cs.meta
