+++ Assets/Scripts/Player/PlayerDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0434c61fe02950f4caffc6f4f837c371
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
