+++ Assets/Scripts/Player/PlayerPowerUps.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 39b082ce02bf7e142863e4fe46eb3d70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
